Ecs (#43)
* ECS steps:
* EnTT 3.2.2 converted to use eastl + allocators  : done
* ask vblanco for permission to use his work: done
* Fix mesh building in mesh_add_surface_from_arrays
* A few signed/unsigned fixups
*  Update EASTL to 3.16.1 + fixups for install
* Some install changes + unified hash map classes.
Replace map with hash in many places.
* HashMap name normalization
* Use HashMap in more places and introduce HashSet.
* Workaround clang>7 glibc>=2.31 bug
* Godot octree was newer used with different allocator, removed that template
parameter.
Renamed/moved NodePath private Data struct to NodePathData.
* Remove builtin physics server, Bullet is the one left.
* Reduce includes in visual server scene header a bit.
Add fast lockless queue
* Remove unused methods from visual_server
* Some struct layout compaction.
* Convert make_binds to vararg template
Rename `call` var-arg helper to call_va.
* Use eastl::function to store requests in the command queue.
* Simplify mt queue and server macros
Now it passes eastl::function instances to the queue, reducing complexity.
* Integrate changes from upstream 3.2.1 rc1.
* Fix installing of concurrent queue lib.

+++ core/pool_vector.h
@@ -393,7 +393,7 @@ public:
     const T & operator[](int p_index) const;
 
     eastl::span<const T> toSpan() const {
-        return { read().ptr(),size()};
+        return { read().ptr(),size_t(size())};
     }
 
     Error resize(int p_size);
